+++ src/main/kotlin/org/jetbrains/plugins/d2/action/exts.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.plugins.d2
+package org.jetbrains.plugins.d2.action
 
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
@@ -7,8 +7,8 @@ import com.intellij.openapi.vfs.VirtualFile
 import org.jetbrains.plugins.d2.editor.D2Viewer
 import org.jetbrains.plugins.d2.file.D2FileType
 
-val VirtualFile.isD2: Boolean
+internal val VirtualFile.isD2: Boolean
   get() = fileType == D2FileType
 
-val AnActionEvent.d2FileEditor
+internal val AnActionEvent.d2FileEditor: D2Viewer
   get() = (getData(PlatformDataKeys.FILE_EDITOR) as TextEditorWithPreview).previewEditor as D2Viewer