+++ src/main/kotlin/org/jetbrains/plugins/d2/editor/D2MessagesAction.kt
@@ -1,10 +1,9 @@
-package com.dvd.intellij.d2.ide.action
+package org.jetbrains.plugins.d2.editor
 
 import com.intellij.execution.impl.ConsoleViewImpl
 import com.intellij.execution.ui.ConsoleViewContentType
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.components.service
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.wm.ToolWindowId
@@ -12,7 +11,7 @@ import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.ui.content.ContentFactory
 import com.intellij.ui.content.MessageView
 import org.jetbrains.plugins.d2.D2Bundle
-import org.jetbrains.plugins.d2.editor.D2Service
+import org.jetbrains.plugins.d2.action.d2FileEditor
 import java.awt.BorderLayout
 import javax.swing.JPanel
 
@@ -37,9 +36,8 @@ private class D2MessagesAction : AnAction(), DumbAware {
     val console = content.component.getComponent(0) as ConsoleViewImpl
     console.clear()
     messageView.contentManager.setSelectedContent(content)
-    for (command in service<D2Service>().map.values) {
-      console.print(command.log.toString(), ConsoleViewContentType.LOG_INFO_OUTPUT)
-    }
+    val state = e.d2FileEditor.renderManager.state ?: return
+    console.print(state.log.toString(), ConsoleViewContentType.LOG_INFO_OUTPUT)
 
     val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(ToolWindowId.MESSAGES_WINDOW)
     if (toolWindow != null && !toolWindow.isActive) {