+++ src/main/kotlin/org/jetbrains/plugins/d2/editor/D2MissingCompilerNotificationProvider.kt
@@ -1,13 +1,14 @@
 package org.jetbrains.plugins.d2.editor
 
 import com.intellij.openapi.fileEditor.FileEditor
+import com.intellij.openapi.fileEditor.TextEditorWithPreview
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.ui.EditorNotificationPanel
 import com.intellij.ui.EditorNotificationProvider
 import org.jetbrains.annotations.Nls
-import org.jetbrains.plugins.d2.isD2
+import org.jetbrains.plugins.d2.action.isD2
 import java.util.function.Function
 import java.util.function.Supplier
 import javax.swing.JComponent
@@ -19,7 +20,7 @@ private class D2MissingCompilerNotificationProvider : EditorNotificationProvider
     }
 
     return Function { fileEditor ->
-      fileEditor.getUserData(D2_FILE_NOTIFICATION)?.let {
+      ((fileEditor as? TextEditorWithPreview)?.previewEditor as? D2Viewer)?.renderManager?.fileNotification?.let {
         D2MissingCompilerNotificationPanel(fileEditor, it)
       }
     }